# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CallCommand AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelList"
    post:
      operationId: createChannel
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      operationId: updateChannel
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      responses:
        "204": { description: Deleted }
  /flows:
    get:
      operationId: listFlows
      tags: [flows]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallFlowList"
    post:
      operationId: createFlow
      tags: [flows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallFlowBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallFlow"
  /flows/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      operationId: getFlow
      tags: [flows]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallFlow"
    patch:
      operationId: updateFlow
      tags: [flows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCallFlowBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallFlow"
    delete:
      operationId: deleteFlow
      tags: [flows]
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Channel:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        phoneNumber: { type: ["string", "null"] }
        type: { type: string }
        defaultRoute: { type: ["string", "null"] }
        isActive: { type: boolean }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [id, userId, name, type, isActive, isDefault, createdAt, updatedAt]
    ChannelList:
      type: array
      items: { $ref: "#/components/schemas/Channel" }
    CreateChannelBody:
      type: object
      properties:
        name: { type: string }
        phoneNumber: { type: ["string", "null"] }
        type: { type: string }
        defaultRoute: { type: ["string", "null"] }
        isActive: { type: boolean }
      required: [name, type]
    UpdateChannelBody:
      type: object
      properties:
        name: { type: string }
        phoneNumber: { type: ["string", "null"] }
        type: { type: string }
        defaultRoute: { type: ["string", "null"] }
        isActive: { type: boolean }
    FlowNode:
      type: object
      properties:
        id: { type: string }
        flowId: { type: string }
        type: { type: string }
        label: { type: ["string", "null"] }
        config: { type: object }
        nextNodeId: { type: ["string", "null"] }
        nextNodeIdFalse: { type: ["string", "null"] }
        orderIndex: { type: integer }
      required: [id, flowId, type, config, orderIndex]
    CallFlow:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        channelId: { type: ["string", "null"] }
        startNodeId: { type: ["string", "null"] }
        isActive: { type: boolean }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/FlowNode" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, userId, name, isActive, nodes, createdAt, updatedAt]
    CallFlowList:
      type: array
      items: { $ref: "#/components/schemas/CallFlow" }
    CreateCallFlowBody:
      type: object
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
        channelId: { type: ["string", "null"] }
        isActive: { type: boolean }
        nodes:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              label: { type: ["string", "null"] }
              config: { type: object }
              nextNodeRef: { type: ["string", "null"] }
              nextNodeFalseRef: { type: ["string", "null"] }
              ref: { type: ["string", "null"] }
            required: [type, config]
      required: [name]
    UpdateCallFlowBody:
      type: object
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
        channelId: { type: ["string", "null"] }
        isActive: { type: boolean }
        nodes:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              label: { type: ["string", "null"] }
              config: { type: object }
              nextNodeRef: { type: ["string", "null"] }
              nextNodeFalseRef: { type: ["string", "null"] }
              ref: { type: ["string", "null"] }
            required: [type, config]
...