# lib/api-zod/src/generated/types/callFlow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CallCommand AI API
 * OpenAPI spec version: 0.1.0
 */
import type { FlowNode } from "./flowNode";

export interface CallFlow {
  id: string;
  userId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  startNodeId?: string | null;
  isActive: boolean;
  nodes: FlowNode[];
  createdAt: Date;
  updatedAt: Date;
}