# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CallCommand AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: string;
  userId: string;
  name: string;
  /** @nullable */
  phoneNumber?: string | null;
  type: string;
  /** @nullable */
  defaultRoute?: string | null;
  isActive: boolean;
  isDefault: boolean;
  createdAt: Date;
  updatedAt: Date;
}