# lib/api-zod/src/generated/types/createCallFlowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CallCommand AI API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCallFlowBodyNodesItem } from "./createCallFlowBodyNodesItem";

export interface CreateCallFlowBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  channelId?: string | null;
  isActive?: boolean;
  nodes?: CreateCallFlowBodyNodesItem[];
}