# lib/api-zod/src/generated/types/updateCallFlowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CallCommand AI API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCallFlowBodyNodesItem } from "./updateCallFlowBodyNodesItem";

export interface UpdateCallFlowBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  channelId?: string | null;
  isActive?: boolean;
  nodes?: UpdateCallFlowBodyNodesItem[];
}